Update version to 0.1.4

+++ vizdet/bboxes.py
@@ -217,7 +217,7 @@ class BBoxes:
                 "The `labels_conf` should be the same lenght as the `boxes_coords`."
             )
 
-        if not isinstance(boxes_coords[0][0], (int, np.int, np.int64)):
+        if not isinstance(boxes_coords[0][0], (int, np.intc, np.int64)):
             raise ValueError("The `boxes_coords` elements should be integers.")
 
         for idx, coords in enumerate(boxes_coords):
